Added basic OpenGL framework for CPU <-> GPU data transfers
The data transfer seems to be/become a major bottleneck of the whole OpenGL
backend. It's quite fast if you have the GL_EXT_texture_integer extension
which is a fairly new one, but the GL_EXT_gpu_shader4 extension seems to be
necessary for handling these integer textures in a fragment shader.

Besides that pushing/pulling to/from the GPU works for U8 and S16 in general.
There is no integration in the encoding/decoding process yet.

The OpenGL testsuite part performs a basic push-pull-compare testcycle for
U8 and S16 data.