[arithdec] Change bit insertion counter to decrement to 0
Instead of reseting the counter to zero and testing if it is 8,
set counter to 8; decrement and test if zero.  This should eliminate
an instruction.

TODO: it is worth investigating if using the dec instruction is
better in this case than using the compiler generated sub $1,%0;
test %0, %0; jcc ...;  The dec instruction partially updates the
registers -- may cause flags stall.