[decoder] refactor the decoder for async changes
Introduce async stages and tasks.  Decoder states are changed
into stages, and each stage (eventually) will be broken into
multiple tasks automatically by SchroAsync.  This is to make
the decoder (and encoder, eventually) much more finely threaded.
Stages are portions of the decoding process that may have dependencies
on other stages of the same picture or other picture.  Tasks are
inherintly parallelizable subportions of a stage, such as motion
rendering a portion of a picture.