decoder: decouple requirement for output_picture at decoder_push time
Previously, when pushing picture data into the decoder, there is an
assertion that an output_picture is avaliable from the output_queue.
This patch modifies the behaviour to nolonger execute the final
combining picture decoding stage unless it has an output picture.
The api state machine now requests a picture if any picture in the
processing queue is missing an output_picture (and the queue is empty).