decoder: implement reorder buffer model to find earliest picture
A reorder buffer allows simple solving of what is the earliest decodable
picture in a sequence.  This occurs by doing an insertion sort when
adding pictures to the reorder buffer based on the ordinality of picture
number.

This patch should also solve all outstanding issues with picture numbers
wrapping: there is a spec modification that defines if one picture is
before another.

This patch maintains behaviour when coded_order is asserted: pictures
are not sorted in the reorder buffer (this is equivalent to having a
ROB size of 1)

An extra bonus, is improved resilience in the event of an errored
sequnce missing a picture: the picture is simply skipped, no fatal
error is flagged.