decoder: pipelined parsing of EOS and the subsequent sequence.
There is a timelag between receiving an EOS at the input to the
decoder and the decoder state producing EOS.  In that time, it
would be nice if a user can continue to push data to the decoder.

This patch implements this.

New API:
 - schro_decoder_begin_instance()
 - schro_decoder_end_instance()

In order to make use of back-to-back sequence parsing:
 - begin_instance() must be called just after an EOS is pushed
   into the decoder.
 - end_instance() must be called when decoder_wait() returns
   SCHRO_DECODER_EOS.
The above is not required when using autoparse_push() or
autoparse_wait().

NB, this patch does not implement any threaded decoding on
the subsequent sequence, only the parsing functions