schroparse: add special startup sync case for ExtraFastSync
Under normal circumstances one data unit + header from next are required
to sync to the bytestream, however, many streams start data unit aligned.
To avoid delays in decoding the first data unit (which may be a seqhdr),
assume that the stream is synced at startup, if it starts with a data
unit and the data unit ends within the first chunk of data passed to the
decoder.