decoder: when field_coding and pulling a frame, remove both pictures
Since the RoB contains two pictures for each frame when field_coding
(and they are in order [1]), the second field picture must also be
removed and discarded at decoder_pull() time.

Handles error case, where second field isn't the pair to the first.
since by definition the RoB has to be full to be considering pulling
a field, don't discard the unrelated second field.  This will stop
a deadlock on the RoB

[1] assuming !coded_order.