decoder: fix obscure motion compensation crash with skipped refs
Consider the following:
  X(0), X(1), B(2)

If picture B(2) is predicted from I(0) and X(1), and one of X(0) or X(1)
is skipped, and B(2) references a block that goes over the edge of one
picture near the end of the frame, then the decoder can segfault.

Actually this can probably occur with single picture references too.

This can be fixed by ensuring that the skipped upconverted reference is
the same size as a non skipped one, with respect to edge extension.

NB, afaict, the decoder is not safe if a block size were to be greater
than the edge extension.  A suggested fix would be to use the single pel
extra slow path to avoid a segfault in these cases.  If it were the case
that the block size became larger than the exension, then a per instance
variable can be used to hold the high-tide block size and use that for
all future extension.