decoder: at EOS, discard dangling fields when field_coding
When field_coding, a frame consists of two fields.  At EOS, a malformed
stream may not have both fields present, in this case, discard the
frame.

Not doing so can risk production of two frames (each with a single
field) at the boundary of two malformed streams. NB, it is easier to
detect this fault at EOS.