decoder: protect against unexpected backwards 'seek'
If a sequence were to suddenly jump backwards (in terms of picture
number) and there were no notification of this discontinuity,
pictures may get stuck in the reorder buffer, causing disturbed
playback.

Detect this case, and issue an EOS.

NB, it isn't required to detect a forwards discontinuity, since this
will naturally flush the reorder buffer.