unpack: improve table lookup decode
Remove the number of symbols that can be decoded for a given
code entry, since it was only used to trigger the slow path,
which can be done other ways.

The last symbol for the last entry (all zeros) is truncated so
we can avoid a conditional on the array index.  The last symbol
is the only one that could theoretically spill over to the next
entry.