lowdelay: Add fast path for decode
This fast path requires that slices_x and slices_y evenly divide
the DC subbands.  This means that every slice has the same layout.
This technique could be expanded to handle multiple layouts.

It is reasonable to assume that efficient encoder implementations
will choose values for slice_x and slice_y that minimize the number
of layouts, and also that these layouts will not change from picture
to picture.